+++ styleguide/components/21-atoms/accolade/accolade.config.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = {
-  hidden: true,
-  context: {
-    position: 'bottom-center'
-  }
-};