+++ styleguide/components/21-atoms/fieldset/fieldset.config.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = {
-  hidden: true,
-  context: {
-    text: 'Lorem ipsum dolor set ilir el.'
-  }
-};