+++ styleguide/components/21-atoms/pubdate/pubdate.config.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-  title: 'Publication date',
-  hidden: true,
-  context: {
-    pubdate: '2017-12-31'
-  }
-};