+++ styleguide/components/31-molecules/quote/quote.config.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  context: {
-    text: 'This is a quote text.'
-  }
-};